requirements.txt
python-dotenv
requests
mysql-connector-python
passlib
pytest
pytest-env
pytest-mock
pylint
pylint-pydantic
pytest-dotenv
wrapt_timeout_decorator
paramiko
fastapi
fastapi_jwt_auth
pydantic[email]
uvicorn[standard]
python-multipart
